moved token creator to new module
