+++ auth/build.gradle
@@ -10,6 +10,7 @@ configurations {
 
 dependencies {
 	implementation project(":core")
+	implementation project(":token")
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
