+++ auth/src/main/java/br/com/gsistemas/auth/AuthApplication.java
@@ -6,6 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
@@ -13,6 +14,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 @EnableConfigurationProperties(value = JwtConfiguration.class)
 @EntityScan({"br.com.gsistemas.core.model"})
 @EnableJpaRepositories({"br.com.gsistemas.core.repository"})
+@ComponentScan("br.com.gsistemas")
 public class AuthApplication {
 
 	public static void main(String[] args) {
