+++ auth/src/main/resources/application.yml
@@ -22,6 +22,7 @@ eureka:
   client:
     service-url:
       defaultZone: http://localhost:8081/eureka/
+    register-with-eureka: true
   instance:
     prefer-ip-address: true
 #    hostname: localhost
