+++ settings.gradle
@@ -5,4 +5,5 @@ include('course')
 include('core')
 include('gateway')
 include('auth')
+include 'token'
