+++ token/build.gradle
@@ -1,0 +1,24 @@
+plugins {
+    id 'org.springframework.boot' version '3.1.2'
+}
+
+configurations {
+    compileOnly {
+        extendsFrom annotationProcessor
+    }
+}
+
+dependencies {
+    implementation project(":core")
+
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+
+    implementation 'com.nimbusds:nimbus-jose-jwt:9.31'
+    implementation 'jakarta.servlet:jakarta.servlet-api'
+
+    compileOnly 'org.projectlombok:lombok'
+
+    annotationProcessor 'org.projectlombok:lombok'
+
+    testImplementation 'org.springframework.security:spring-security-test'
+}
